split up paste handling according to source
